Clean up exports, update README
